+++ mackup.py
@@ -152,6 +152,10 @@ SUPPORTED_APPS = {
 
     'Nano': ['.nanorc'],
 
+    'nvALT': [PREFERENCES + 'net.elasticthreads.nv.plist',
+              APP_SUPPORT + 'Notational Velocity',
+              APP_SUPPORT + 'Notational Data'],
+
     'Oh My Zsh': ['.oh-my-zsh'],
 
     'PCKeyboardHack': [PREFERENCES + 'org.pqrs.PCKeyboardHack.plist'],