+++ mackup.py
@@ -329,7 +329,7 @@ SUPPORTED_APPS = {
 
 
 # Current version
-VERSION = '0.4.2'
+VERSION = '0.4.3'
 
 # Mode used to backup files to Dropbox
 BACKUP_MODE = 'backup'